Created NewQuestionPage and refactored components to make the code cleaner
