+++ ui/src/message_popup/author/new_question_page.jsx
@@ -1,0 +1,14 @@
+/* @flow weak */
+import React from 'react';
+
+import QuestionPage from './question_page.jsx';
+
+export default React.createClass({
+  displayName: 'NewQuestionPage',
+
+  render(){
+    return (
+      <QuestionPage />
+      );
+  }
+});
