+++ ui/src/message_popup/author/question_editing_components/editing_component.jsx
@@ -1,0 +1,14 @@
+/* @flow weak */
+import QuestionText from './question_text.jsx';
+import Students from './students.jsx';
+import Examples from './examples.jsx';
+import LearningObjectives from './learning_objectives.jsx';
+import Indicators from './indicators.jsx';
+
+export {
+  QuestionText,
+  Students,
+  Examples,
+  LearningObjectives,
+  Indicators
+};
